Make numeric widening explicit

+++ modules/client/src/main/scala/rx/redis/serialization/ByteBufAccess.scala
@@ -73,7 +73,7 @@ object ByteBufAccess extends BytesAccess[ByteBuf] {
     a.toString(charset)
 
   def writeByte(a: ByteBuf, b: Byte): Unit = {
-    a.writeByte(b)
+    a.writeByte(b.toInt)
   }
 
   def writeBytes(a: ByteBuf, bs: Array[Byte]): Unit = {
